Separating resource tagging
File that will replace the PS inline script to tag resources. More resilient in that it will retry in the case of an error (though was not sure what type of exception the transient error was). 

Script requires 3 parameters - resourceGroupName, tagId, and deploymentId. All resources in resourceGroupName will be tagged with 'tagId = deploymentId'