Tagresources (#8)
* Separating resource tagging

File that will replace the PS inline script to tag resources. More resilient in that it will retry in the case of an error (though was not sure what type of exception the transient error was). 

Script requires 3 parameters - resourceGroupName, tagId, and deploymentId. All resources in resourceGroupName will be tagged with 'tagId = deploymentId'

* Utilize new tagging script

Replace the inline powershell script that does resource tagging with script that does tagging (and is more resilient in that it will retry a set of the resource tags).

* Wait on set exception

Wait 5 seconds on a failed set of a resource. It is a timing issue (most likely) so give it a little time and see if that helps.